requirements.txt
flask
kronoslabs
python-dotenv
flask-cors